// src/generated/types.ts
export interface CharacterSkillResponse {
  skill_id: number;
  skill_name: string;
  group_id: number;
  group_name: string;
  trained_skill_level: number;
  active_skill_level: number;
  skillpoints_in_skill: number;
  is_in_queue: boolean;
  queue_level?: number | null;
  is_injected: boolean;
}
// ...
export interface CharacterSkillsResponse {
  character_id: number;
  skills: CharacterSkillResponse[];
  groups: SkillGroupResponse[];
}
// ...
export interface SkillGroupResponse {
  group_id: number;
  group_name: string;
  total_levels: number;
  trained_levels: number;
  has_trained_skills: boolean;
}